+++ package.json
@@ -7,9 +7,13 @@
     "build": "next build",
     "start": "next start",
     "export": "next build && next export && next-image-export-optimizer && touch out/.nojekyll",
-    "deploy":"next build && next export && next-image-export-optimizer && touch out/.nojekyll && git add out/ && git commit -m \"Deploy gh-pages\" && git subtree push --prefix out origin gh-pages"
+    "deploy": "next build && next export && next-image-export-optimizer && touch out/.nojekyll && git add out/ && git commit -m \"Deploy gh-pages\" && git subtree push --prefix out origin gh-pages"
   },
   "dependencies": {
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.9.2",
     "image-trace-loader": "^1.0.2",
     "imagemin-svgo": "^9.0.0",
     "next": "^12.1.6",
