+++ styles/Nav.module.scss
@@ -46,15 +46,14 @@
 }
 
 .nav {
-    background: $primary;
+    background: $secondary;
     color: $white;
-    height: 3rem;
-    min-height: 3rem;
-    // width: 100vw;
-    padding: 10px;
+    height: 7rem;
+    min-height: 5rem;
+    width: 100%;
     top: 0;
     @include flex-row;
-    @include left-right-centered;
+    @include xcross-centered;
     @include sticky;
     @include unselectable;
 
@@ -84,20 +83,67 @@
         cursor: pointer;
     }
 
+    // search bar
+    .inputcontainer {
+        display: flex;
+        width: 50%;
+        min-height: 100%;
+        align-items: center;
+    }
+
+    .searchbar {
+        box-sizing: border-box;
+        flex-direction: row;
+        align-items: center;
+        font-size: 16px;
+        border: 2px solid #E7E7E7;
+        box-shadow: 0px 5px 10px rgba(131, 131, 131, 0.08);
+        border-radius: 40px;
+        width: 100%;
+        min-height: 100%;
+    }
+    .searchicon {
+        color: $primary;
+        text-align: center;
+        margin-right: 3%;
+    }
+
+    .signinbutton {
+        box-sizing: border-box;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        padding: 0px;
+        width: 100px;
+        height: 42px;
+        color: $primary;
+        background-color: $secondary;
+        border: 1px solid $primary;
+        border-radius: 66px;
+        font-size: 15px;
+        line-height: 26px;
+        font-weight: 500;
+        letter-spacing: 0.46px;
+        text-transform: uppercase;
+    }
+
+    .innercontainer {
+        @include flex-row;
+        @include left-right-centered-even;
+        width: 60%;
+        flex-grow: 1s;
+        min-height: 100%;
+    }
+
     @media only screen and (max-width: 800px) {
         justify-content: space-between;
         @include xcross-centered;
 
         .clogo {
             flex: 1;
-            margin-right: 2.5rem;
             @include flex-column;
             @include cross-centered;
-        }
-
-        .hamburgermenu {
-            position: absolute;
-            display: block;
         }
 
         .list {
@@ -111,73 +157,4 @@
         }
     }
 
-    .hamburgermenu {
-        --stroke-width: 0.25rem;
-        box-sizing: content-box;
-        color: hsla(0,0%,100%,.74902);
-        color: var(--text-color--secondary);
-        background-color: transparent;
-        cursor: pointer;
-        height: 1.25rem;
-        line-height: 0;
-        padding: 1rem;
-        position: relative;
-        width: 1.5rem;
-        border: 0;
-        @include hover() {
-            color: $link-light-hover-color;
-            text-decoration: $link-light-hover-color;
-        }
-        
-        .hamburgermenuicon {
-            height: 100%;
-            position: relative;
-            width: 100%;
-            &.isactive {
-                .hamburgermenuline {
-                    &:first-child {
-                        width: 100%;
-                        --rotate: 45deg;
-                        top: 50%;
-                    }
-                    &:nth-child(2) {
-                        width: 100%;
-                        --rotate: 45deg;
-                    }
-                    &:nth-child(3) {
-                        width: 100%;
-                        --rotate: -45deg;
-                        bottom: 50%;
-                    }
-                }
-            }
-            .hamburgermenuline {
-                --duration: 0.2s;
-                --timing: ease-out;
-                background-color: currentColor;
-                border-radius: 100rem;
-                display: block;
-                height: var(--stroke-width);
-                position: absolute;
-                transform: translateY(0) rotate(0deg);
-                transform: translateY(var(--translateY,0)) rotate(var(--rotate,0deg));
-                width: 100%;
-                transition: bottom calc(var(--duration)*0.5) var(--timing) calc(var(--duration)*0.5),color calc(var(--duration)*0.5) var(--timing),top calc(var(--duration)*0.5) var(--timing) calc(var(--duration)*0.5),transform calc(var(--duration)*0.5) var(--timing),width calc(var(--duration)*0.5) var(--timing) calc(var(--duration)*0.5);
-                &:first-child {
-                    --translateY: -50%;
-                    top: calc(var(--stroke-width)*0.5);
-                }
-                &:nth-child(2) {
-                    width: 75%;
-                    --translateY: -50%;
-                    top: 50%;
-                }
-                &:nth-child(3) {
-                    width: 50%;
-                    --translateY: 50%;
-                    bottom: calc(var(--stroke-width)*0.5)
-                }
-            }
-        }
-    }
 }
