+++ styles/_variables.scss
@@ -7,7 +7,7 @@
 
 // Theme colors
 
-$primary: #0e0e0e !default;
+$primary: #A53692 !default;
 $primary-light: lighten($primary, 75%) !default;
 $secondary: #fff !default;
 $success: #3772FF !default;
