+++ styles/globals.scss
@@ -8,7 +8,7 @@ html,
 body {
   padding: 0;
   margin: 0;
-  overflow-x: hidden;
+  overflow: visible;
   font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
     Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 
@@ -90,7 +90,7 @@ body {
   }
 
   ::-webkit-scrollbar {
-    width: 1rem;
+    display:none
   }
 
   ::-webkit-scrollbar-track {
